+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+- In/Out Garrison Sensemaker now does classification rollup correctly
 
 ## [0.19.10] - 2026-03-16