+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+- Fix Existing Incursions excluding their current classification when doing classification rollup
 
 ## [0.19.10] - 2026-03-16