+++ CHANGELOG.md
@@ -9,6 +9,14 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+
+## [0.19.11] - 2026-03-18
+
+### Added
+
+### Changed
+
+### Fixed
 - In/Out Garrison Sensemaker now does classification rollup correctly
 - Fix Existing Incursions excluding their current classification when doing classification rollup
 - Using all attributes to determine the acm of a resolution relationship